+++ test/e2e/operator.go
@@ -5,12 +5,16 @@ package e2e
 
 import (
 	"context"
+	"fmt"
+	"regexp"
+	"strings"
 	"time"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	"github.com/ugorji/go/codec"
+	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -22,6 +26,45 @@ import (
 	"github.com/Azure/ARO-RP/pkg/operator/controllers/monitoring"
 	"github.com/Azure/ARO-RP/pkg/util/ready"
 )
+
+func updatedObjects(ctx context.Context, nsfilter string) ([]string, error) {
+	pods, err := clients.Kubernetes.CoreV1().Pods("openshift-azure-operator").List(ctx, metav1.ListOptions{
+		LabelSelector: "app=aro-operator-master",
+	})
+	if err != nil {
+		return nil, err
+	}
+	if len(pods.Items) != 1 {
+		return nil, fmt.Errorf("%d aro-operator-master pods found", len(pods.Items))
+	}
+	b, err := clients.Kubernetes.CoreV1().Pods("openshift-azure-operator").GetLogs(pods.Items[0].Name, &corev1.PodLogOptions{}).DoRaw(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	rx := regexp.MustCompile(`msg="(Update|Create) ([-a-zA-Z/.]+)`)
+	changes := rx.FindAllStringSubmatch(string(b), -1)
+	result := make([]string, 0, len(changes))
+	for _, change := range changes {
+		if nsfilter == "" || strings.Contains(change[2], "/"+nsfilter+"/") {
+			result = append(result, change[1]+" "+change[2])
+		}
+	}
+
+	return result, nil
+}
+
+func dumpEvents(ctx context.Context, namespace string) error {
+	events, err := clients.Kubernetes.EventsV1().Events(namespace).List(ctx, metav1.ListOptions{})
+	if err != nil {
+		return err
+	}
+
+	for _, event := range events.Items {
+		log.Debugf("%s. %s. %s", event.Action, event.Reason, event.Note)
+	}
+	return nil
+}
 
 var _ = Describe("ARO Operator - Internet checking", func() {
 	var originalURLs []string
@@ -105,6 +148,8 @@ var _ = Describe("ARO Operator - Geneva Logging", func() {
 			err := dumpEvents(context.Background(), "openshift-azure-monitoring")
 			Expect(err).NotTo(HaveOccurred())
 		}
+		Expect(err).NotTo(HaveOccurred())
+
 		initial, err := updatedObjects(context.Background(), "openshift-azure-logging")
 		Expect(err).NotTo(HaveOccurred())
